refactor: lift ConstrDispatcher rules into SumDispatch (Phase 3a)
Moves `shouldDelegateToUplcConstr`'s four rules and `dispatchNil` from
`typegens/ConstrDispatcher.scala` into `SumDispatch.chooseConstrOutputRepr`
and `SumDispatch.dispatchNil`. `Lowering.lowerSIR`'s non-Nil Constr path
routes through `SumDispatch.genConstr` / `ProdDispatch.genConstr`, which
call `chooseConstrOutputRepr` at the dispatcher entry. The Nil path stays
inline (the dispatcher returns the typegen directly), but `dispatchNil`'s
`Some(targetType)` branch now also honors `inUplcConstrListScope`. Without
that, Nils inside an UplcConstr-scoped intrinsic body whose surrounding
if-then-else's LUB-stripped target type lost its `@UplcRepr(UplcConstr)`
annotation would emerge as `SumBuiltinList`, desyncing from match-branch
`SumUplcConstr` reprs and triggering a `uplcConstrToBuiltinList` conversion
against an unresolved TypeVar element.

Removes the typegen-internal `ConstrDispatcher.shouldDelegateToUplcConstr`
calls from `SumCaseSirTypeGenerator`, `SumListCommonSirTypeGenerator`, and
`ProductCaseSirTypeGenerator`. Deletes the now-orphan `ConstrDispatcher.scala`.

`jvm/test`: 432 passing.
