+++ scalus-core/shared/src/main/scala/scalus/compiler/sir/lowering/typegens/SumListCommonSirTypeGenerator.scala
@@ -102,11 +102,8 @@ trait SumListCommonSirTypeGenerator extends SirTypeUplcGenerator {
         optTargetType: Option[SIRType]
     )(using lctx: LoweringContext): LoweredValue = {
         import SumListCommonSirTypeGenerator.*
-        // Context-driven delegation (see ConstrDispatcher).
-        ConstrDispatcher.shouldDelegateToUplcConstr(constr, loweredArgs, optTargetType) match
-            case Some(other) if other ne this =>
-                return other.genConstrLowered(constr, loweredArgs, optTargetType)
-            case _ =>
+        // Routing decisions live in SumDispatch.chooseConstrOutputRepr now (see
+        // ConstrDispatcher.shouldDelegateToUplcConstr for the original rules).
         constr.name match
             case SIRType.List.NilConstr.name | SIRType.BuiltinList.Nil.name | PairNilName =>
                 genNil(constr.tp, constr.anns.pos)
